MP-46 have the login redirect you to profile page instead
