+++ src/pages/LandingPage.jsx
@@ -5,19 +5,8 @@ import Footer from '../components/general/Footer'
 import { FaqWithImage } from '../components/landingpage/FAQ'
 import { EmailBanner } from '../components/landingpage/SubscribeToNews'
 // import NotFoundImage from '../pages/PageNotFound'
-import { useLocation } from 'react-router-dom'
-import { notifications } from '@mantine/notifications'
 
 export default function LandingPage() {
-  const location = useLocation()
-  if(location.state){
-    location.state.loginSuccess && notifications.show({
-      title: 'Login Success',
-      message: location.state.message,
-      color: 'green',
-      autoClose: 5000
-    })
-  }
   return (
     <>
       <MinimalNavbar/>
